# Documentation/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Gradox2 API",
        "description": "API de gestión de archivos académicos para la aplicación Gradox2",
        "contact": {
            "name": "Gradox2 Team",
            "email": "gradox2app@gmail.com"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.html"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/files/{id}/comments/{commentId}": {
            "put": {
                "tags": [
                    "Hilos / Comentarios"
                ],
                "summary": "Editar comentario",
                "description": "Edita el contenido de un comentario propio (marca editedAt)",
                "operationId": "editComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID del comentario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EditCommentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Comentario actualizado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "No eres el autor del comentario"
                    },
                    "404": {
                        "description": "Archivo o comentario no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hilos / Comentarios"
                ],
                "summary": "Eliminar comentario",
                "description": "Elimina un comentario (autor o MASTER). Si era el último, se elimina el hilo",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "commentId",
                        "in": "path",
                        "description": "ID del comentario",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comentario eliminado"
                    },
                    "400": {
                        "description": "Sin permiso para eliminar"
                    },
                    "404": {
                        "description": "Archivo o comentario no encontrado"
                    }
                }
            }
        },
...
        "/files/{id}/thread/lock": {
            "post": {
                "tags": [
                    "Hilos / Comentarios"
                ],
                "summary": "Bloquear hilo",
                "description": "Bloquea el hilo del archivo (uploader o MASTER): no se admiten nuevos comentarios",
                "operationId": "lockThread",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hilo bloqueado"
                    },
                    "400": {
                        "description": "Sin permiso para bloquear"
                    },
                    "404": {
                        "description": "Archivo o hilo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Hilos / Comentarios"
                ],
                "summary": "Desbloquear hilo",
                "description": "Desbloquea el hilo del archivo (uploader o MASTER)",
                "operationId": "unlockThread",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hilo desbloqueado"
                    },
                    "400": {
                        "description": "Sin permiso para desbloquear"
                    },
                    "404": {
                        "description": "Archivo o hilo no encontrado"
                    }
                }
            }
        },
        "/files/{id}/comments": {
            "get": {
                "tags": [
                    "Hilos / Comentarios"
                ],
                "summary": "Listar comentarios",
                "description": "Lista los comentarios del hilo de un archivo (paginado, visibilidad del archivo aplicada)",
                "operationId": "getThreadComments",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0,
                            "minimum": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 20,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista devuelta",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CommentResponse"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Archivo no encontrado o sin acceso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Hilos / Comentarios"
                ],
                "summary": "Crear comentario",
                "description": "Publica un comentario en el hilo del archivo (crea el hilo si no existe). Puede responder a otro comentario y/o referenciar un archivo",
                "operationId": "createComment",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del archivo",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCommentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Comentario publicado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos, hilo bloqueado o padre no válido"
                    },
                    "404": {
                        "description": "Archivo no encontrado o sin acceso"
                    }
                }
            }
        },
...
        "/subjects/{id}": {
            "get": {
                "tags": [
                    "Asignaturas"
                ],
                "summary": "Ver asignatura",
                "description": "Devuelve el detalle de una asignatura con su curso",
                "operationId": "getSubject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Asignatura devuelta",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubjectResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Asignatura no encontrada",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubjectResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EditCommentRequest": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 0
                    }
                }
            },
            "CommentResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "content": {
                        "type": "string"
                    },
                    "authorUsername": {
                        "type": "string"
                    },
                    "parentCommentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "referencedFile": {
                        "$ref": "#/components/schemas/FileReference"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "editedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "FileReference": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "fileType": {
                        "type": "string"
                    },
                    "subject": {
                        "type": "string"
                    },
                    "available": {
                        "type": "boolean"
                    }
                }
            },
...
            "CreateCommentRequest": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 0
                    },
                    "referencedFileId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "parentCommentId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "SubjectResponse": {
                "type": "object",
                "description": "Respuesta con los datos de una asignatura",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID de la asignatura",
                        "example": 1
                    },
                    "code": {
                        "type": "string",
                        "description": "Código de la asignatura",
                        "example": "MAT1"
                    },
                    "name": {
                        "type": "string",
                        "description": "Nombre de la asignatura",
                        "example": "Matemáticas I"
                    },
                    "courseId": {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID del curso al que pertenece",
                        "example": 1
                    },
                    "courseCode": {
                        "type": "string",
                        "description": "Código del curso",
                        "example": 1
                    },
                    "courseName": {
                        "type": "string",
                        "description": "Nombre del curso",
                        "example": "Primer curso"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}